Add tests for dictionaryValue and arrayValue
